src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadow-hierarchy-exception-expected.txt

CONSOLE ERROR: line 12: Uncaught TypeError: span.createShadowRoot is not a function
Appending an ancestor to a shadow tree should throw an exception

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
